Resultado dos experimentos

+++ trab4/boosting_AdaBoostClassifier.py
@@ -31,6 +31,6 @@ def run(num=2, lr=1):
 	
 if __name__ == "__main__":
 	print(sys.argv)
-	run(int(sys.argv[1]), int(sys.argv[2]))
+	run(int(sys.argv[1]), float(sys.argv[2]))
